+++ tools/quake2/ericw-tools/src/include/qbsp/writebsp.hh
@@ -21,8 +21,10 @@
 
 #pragma once
 
-size_t ExportMapTexinfo(size_t texinfonum);
-size_t ExportMapPlane(size_t planenum);
+#include <cstdint>
+
+int32_t ExportMapTexinfo(size_t texinfonum);
+int32_t ExportMapPlane(size_t planenum);
 
 void BeginBSPFile();
 void FinishBSPFile();
